Fix bad optimization of ORDER BY in subquery if it contains WITH FILL
