+++ tests/queries/0_stateless/2015_order_by_with_fill_misoptimization.reference
@@ -1,0 +1,9 @@
+
+
+
+
+
+
+
+
+Hello
