+++ tests/queries/0_stateless/2015_order_by_with_fill_misoptimization.sql
@@ -1,0 +1,1 @@
+SELECT s FROM (SELECT 5 AS x, 'Hello' AS s ORDER BY x WITH FILL FROM 1 TO 10) ORDER BY s;
